Enusre public header and generated code have no implicit converion.

+++ src/google/protobuf/stubs/hash.h
@@ -348,7 +348,7 @@ struct hash<const char*> {
   inline size_t operator()(const char* str) const {
     size_t result = 0;
     for (; *str != '\0'; str++) {
-      result = 5 * result + *str;
+      result = 5 * result + static_cast<size_t>(*str);
     }
     return result;
   }
